AOC2017.Util.Accum

Documentation

class (Monoid w, Monad m) => MonadAccum w m | m -> w where

Minimal complete definition

add, look

Methods

add :: w -> m ()

look :: Monoid w => m w

newtype AccumT w (m :: * -> *) a :: * -> (* -> *) -> * -> *

An accumulation monad parameterized by:

w - the output to accumulate.

m - The inner monad.

The return function produces the output mempty, while >>= combines the outputs of the subcomputations using mappend.

This monad transformer is similar to both state and writer monad transformers. Thus it can be seen as

a restricted append-only version of a state monad transformer or

a writer monad transformer with the extra ability to read all previous output.

Constructors

AccumT (w -> m (a, w))

execAccumT :: Monad m => AccumT w m a -> w -> m w

Extract the output from an accumulation computation.

execAccumT m w = liftM snd (runAccumT m w)

evalAccumT :: (Monad m, Monoid w) => AccumT w m a -> w -> m a

Evaluate an accumulation computation with the given initial output history and return the final value, discarding the final output.

evalAccumT m w = liftM fst (runAccumT m w)

type Accum w = AccumT w Identity

An accumulation monad parameterized by the type w of output to accumulate.

The return function produces the output mempty, while >>= combines the outputs of the subcomputations using mappend.

accum :: Monad m => (w -> (a, w)) -> AccumT w m a

Construct an accumulation computation from a (result, output) pair. (The inverse of runAccum.)

runAccum :: Accum w a -> w -> (a, w)

Unwrap an accumulation computation as a (result, output) pair. (The inverse of accum.)

execAccum :: Accum w a -> w -> w

Extract the output from an accumulation computation.

execAccum m w = snd (runAccum m w)

evalAccum :: Monoid w => Accum w a -> w -> a

Evaluate an accumulation computation with the given initial output history and return the final value, discarding the final output.

evalAccum m w = fst (runAccum m w)

Orphan instances

(Monoid a, Monoid w, MonadWriter w m) => MonadWriter w (AccumT a m)

Methods

writer :: (a, w) -> AccumT a m a

tell :: w -> AccumT a m ()

listen :: AccumT a m a -> AccumT a m (a, w)

pass :: AccumT a m (a, w -> w) -> AccumT a m a