Day10

Description

Documentation

reverseRange :: (IArray a e, Ix i, Num i) => a i e -> (i, i) -> a i e

Reverse elements of an array in a range of indices. The range may wrap.

knotRanges :: (Ix i, Num i) => (i, i) -> [Int] -> [(i, i)]

Given array bounds and a list of lengths, returns a list of ranges in the array, with each one starting at an increasing distance from the end of the previous, wrapping around the ends of the array.

hash :: (IArray a e, Ix i, Num i) => a i e -> [Int] -> a i e

Sequentially reverses all ranges in an array from a list of lengths.

hashString :: String -> [Word8]

Deriving a key from a string by using its codepoints plus some magic numbers, hash [0..255] 64 times, then xor together each group of 16.

day10a :: Int -> String -> Int

day10b :: String -> String