Day11

Description

Documentation

data Step

A single step on a hexagonal grid.

Constructors

N

NE

SE

S

SW

NW

data Pos

A position on a hexagonal grid, in trapezoidal coordinates.

Constructors

Pos
The distance along the N/S axis is given by x + y.

Fields

x :: !Int

The distance along the NE/SW axis.

y :: !Int

The distance along the NW/SE axis.

data Chart

A representation of all hexagons touched along a path.

Constructors

Chart

Fields

bounds :: ((Int, Int), (Int, Int))

Cartesian bounding box.

maxDistance :: Int

The maximum distance from origin.

path :: [((Int, Int), Int)]

Hexagons and distances from origin.

parse :: String -> [Step]

Parse comma-separated steps.

step :: Pos -> Step -> Pos

Move from a position by one step.

walk :: Pos -> Int

Returns the minimum number of steps required to reach the origin.

day11a :: String -> Int

day11b :: String -> Int

chart :: String -> Chart