Day15

Description

Documentation

parse :: String -> (#Word#, Word##)

Returns the initial values for generators A and B.

genA :: Word# -> Word#

One step of generator A.

genB :: Word# -> Word#

One step of generator B.

gen :: (#Word#, Word##) -> (#Word#, Word##)

One step of both generators A and B.

gen' :: (#Word#, Word##) -> (#Word#, Word##)

Step both generators A and B until a multiple of 4 and 8 respectively.

eq16 :: (#Word#, Word##) -> Bool

Tests whether the lower 16 bits of each pair item are equal.

count :: Int -> Int -> (#Word#, Word##) -> Int

count n 0 x counts how many match eq in take n [gen x, gen (gen x), ..]

count' :: Int -> Int -> (#Word#, Word##) -> Int

count' n 0 x counts how many match eq in take n [gen x, gen (gen x), ..]

day15a :: String -> Int

day15b :: String -> Int