Day16

Description

Documentation

data Move i e

A dance move.

Constructors

Spin i

Exchange i i

Partner e e

words' :: String -> [String]

Split words by whitespace or comma.

parse :: String -> [Move Int Char]

Parse a list of dance moves.

(-*-) :: (IArray a e, IArray a' i, Ix i) => a i e -> a' i i -> a i e

Compose two permutations.

(-^-) :: (IArray a i, Ix i) => a i i -> Int -> a i i

Exponentiate a permutation.

exchange :: Eq a => a -> a -> a -> a

Like id, but swapping the two given values.

permuteNames :: IArray a Char => Int -> [Move i Char] -> a Char Char

Given a list of dance moves, map out the resulting partner changes.

permuteIndices :: IArray a Int => Int -> [Move Int e] -> a Int Int

Given a list of dance moves, map out the resulting spins and exchanges.

day16a :: Int -> String -> String

day16b :: Int -> Int -> String -> String