Day18

Description

Documentation

data Ins reg val

A single instruction.

Either represents either a register reference or an immediate value.

Constructors

Rcv reg

Snd (Either reg val)

Set reg (Either reg val)

Add reg (Either reg val)

Mul reg (Either reg val)

Mod reg (Either reg val)

Jgz (Either reg val) (Either reg val)

data MachineSpec m pc reg val

A specification for how to run a machine.

Constructors

MachineSpec

Fields

program :: Array pc (Ins reg val)

The list of instructions.

send :: val -> m ()

Handler for Snd.

recv :: val -> m val

Handler for Rcv.

data MachineState pc reg val

The current state of a machine.

Constructors

MachineState
The machine is running.

Fields

pc :: pc

The program counter.

regs :: Map reg val

The registers.

MachineTerminated
The machine is stopped.

parse :: Read a => String -> [Ins String a]

Parse an assembly listing to instructions.

step :: (Monad m, Ix pc, Num pc, Ord reg, Integral val, Ord val) => MachineSpec m pc reg val -> MachineState pc reg val -> m (MachineState pc reg val)

Evaluate a single instruction.

loop :: (Monad m, Ix pc, Num pc, Ord reg, Integral val, Ord val) => MachineSpec m pc reg val -> MachineState pc reg val -> m ()

Run a machine until its state reaches MachineTerminated.

day18a :: String -> Int64

day18b :: String -> IO Int