Day19

Description

Documentation

data Direction

Represents a cardinal direction.

Constructors

U

D

L

R

parse :: IArray a Char => String -> (a (Int, Int) Char, (Int, Int))

Returns the 2D representation of a string by lines and the index of the only | pipe character on the first line.

rot180 :: Direction -> Direction

Rotates a cardinal direction by 180 degrees.

move :: Num a => (a, a) -> Direction -> (a, a)

Moves a point one step in a cardinal direction.

joints :: (IArray a Char, Ix i, Num i) => a (i, i) Char -> (i, i) -> [Direction]

Returns the cardinal directions around a point which contain non-space.

walk :: (IArray a Char, Ix i, Num i) => a (i, i) Char -> (i, i) -> Direction -> String

Walks a maze from a starting point in a direction until it runs of non-space characters to follow. Returns a list of all characters on the walk.

day19a :: String -> String

day19b :: String -> Int