Day20

Description

Documentation

data Vec3 a

A 3D vector.

Constructors

Vec3

Fields

x :: !a

y :: !a

z :: !a

data Point a

A point with position, velocity, and acceleration in 3D space.

Constructors

Point

Fields

pos :: !(Vec3 a)

vel :: !(Vec3 a)

acc :: !(Vec3 a)

readMany :: Read a => String -> [a]

parse :: Read a => String -> [Point a]

Reads points.

(*+*) :: Num a => Vec3 a -> Vec3 a -> Vec3 a infixl 3

Pointwise addition.

(*-*) :: Num a => Vec3 a -> Vec3 a -> Vec3 a infixl 3

Pointwise subtraction.

step :: Num a => Point a -> Point a

Performs one timestep of velocity and position updates.

manhattan :: Num a => Vec3 a -> a

Returns the Manhattan distance of a vector to the origin.

signumsMatch :: (Eq a, Num a) => Point a -> Bool

Returns true if repeated step will not change the sign of any component of velocity and position.

collide :: Ord a => (b -> a) -> [b] -> [b]

Filters out all elements which are duplicated under transformation.

day20a :: String -> Int

day20b :: String -> [Int]