Day21

Description

Documentation

transforms :: [[[a]] -> [[a]]]

8 affine transformations.

parse :: String -> Map [[Bool]] [[Bool]]

Parse a list of enhancements.

start :: [[Bool]]

The glider.

step :: Ord a => Map [[a]] [[a]] -> [[a]] -> [[a]]

Expand sub-squares according to the rules map.

canonicalize :: Ord a => [[a]] -> [[a]]

Returns the same valeu for all transformations of the same block.

step3

day21 :: Int -> String -> Int

day21 n input returns the number of bits set in the n-th step of transforming the start glider using parse rules from input.

day21a :: String -> Int

day21b :: String -> Int