Day23

Description

Documentation

data Ins reg imm

A single instruction.

Either represents either a register reference or an immediate value.

Constructors

Set

Fields

reg :: reg

val :: Either reg imm

Sub

Fields

reg :: reg

val :: Either reg imm

Mul

Fields

reg :: reg

val :: Either reg imm

Jnz

Fields

cnd :: Either reg imm

jmp :: Either reg imm

data State a reg imm

The current state of a machine.

Constructors

State

Fields

pc :: Int

regs :: a reg imm

parseVal :: Read imm => String -> Either Char imm

Parses an immediate value or a single character register name.

parse :: Read imm => String -> [Ins Char imm]

Parses an assembly listing to instructions.

step :: (IArray a imm, Ix reg, Integral imm) => [Ins reg imm] -> State a reg imm -> Maybe (State a reg imm)

Evaluates a single instruction.

stepOptimized :: (IArray a imm, Ix reg, Integral imm) => [Ins reg imm] -> State a reg imm -> Maybe (State a reg imm)

Evaluates the f = 0 if !isPrime(b) sequence, or a single instruction.

iterateMaybe :: (a -> Maybe a) -> Maybe a -> [a]

Iterates a function until Nothing.

day23a :: String -> Int

day23b :: String -> Int