Day24

Description

Documentation

parse :: (Ord a, Read a) => String -> Map a (Map Int a)

Bidrectionally associates a => n => b and b => n => a, where n is an unique identifier for each a/b line.

deleteOrNull :: Ord k => k -> Map k v -> Maybe (Map k v)

Deletes a key from a Map, returning Nothing if it becomes empty.

buildBridge :: (Num a, Ord a, Ord b, Semigroup s) => (a -> a -> s -> s) -> s -> a -> Map a (Map b a) -> s

Explores all bridges.

day24a :: String -> Int

day24b :: String -> Int