Day25

Description

Documentation

data Move

A movement of the cursor of the Turing machine on its tape.

Constructors

MoveLeft

MoveRight

data State

The state of the the Turing machine.

Constructors

A

B

C

D

E

F

data Op state value

An operation to be performed by the Turing machine.

Constructors

Op

Fields

write :: value

move :: Move

next :: state

data Program state value

The setup for a run of a Turing machine.

Constructors

Program

Fields

start :: state

Initial state of the Turing machine.

steps :: Int

Number of cycles to run the Turing machine. | State transitions.

transitions :: Map (state, value) (Op state value)

data Tape a

A cursor view of a Turing machine's tape.

Constructors

Tape

Fields

before :: [a]

cur :: a

after :: [a]

parseProgram :: (Ord state, Read state, Ord value, Read value) => ReadP (Program state value)

Parses a description to a Program.

writeMoveTape :: Num a => a -> Move -> Tape a -> Tape a

Writes a value then moves the cursor.

step :: (Ord state, Num value, Ord value) => Map (state, value) (Op state value) -> (state, Tape value) -> (state, Tape value)

Performs a single state transition.

day25 :: String -> Int