Day5

Description

Documentation

steps

Arguments

:: (Int -> Int)
After each step, mutate the previous item

-> [Int]
Start at index 0

-> Int
Number of steps until the index is out of bounds

At each step, the current item is used as a relative index shift.

day5a :: String -> Int

day5b :: String -> Int