Day6

Description

Documentation

parse :: (IArray a e, Ix i, Num i, Read e) => String -> a i e

Reads the words of a string to an array.

top :: (IArray a e, Ix i, Ord e) => a i e -> (i, e)

Returns the index and value of first maximum in an array.

redistribute :: (IArray a i, Integral i, Ix i) => a i i -> a i i

Removes the first maximum and redistributes its value throughout the array.

indexDup :: Ord a => [a] -> Maybe (Int, Int)

The indices of the first two duplicated elements in a list.

day6a :: String -> Maybe Int

day6b :: String -> Maybe Int