Day7

Description

Documentation

parseTree :: String -> Maybe [(String, (Int, [String]))]

Parses a string as a table of node name to node weight and children.

findRoot :: Ord a => [(a, (b, [a]))] -> Maybe a

Finds a node with no parents.

mode :: Eq a => [a] -> Maybe a

Returns the most common element.

weighTree :: (Monad m, MonadWriter (First b) m, Ord a, Eq b, Num b) => Map a (b, [a]) -> a -> m b

Returns the total weight of the tree rooted at the given node.

As a side effect, also tell the corrected weight for nodes whose tree weight does not equal the majority of their siblings'.

day7a :: String -> Maybe String

day7b :: String -> Maybe Int