Day8

Description

Documentation

data Instruction reg val

A conditional register instruction.

Constructors

Instruction

Fields

target :: reg

The register to be manipulated.

op :: val -> val

The operation to apply to target.

cond :: val -> Bool

The condition on this instruction.

when :: reg

The register for cond evaluation.

parse :: String -> [Instruction String Int]

Parses a string into a sequence of instructions.

initialize :: Ord reg => val -> [Instruction reg val] -> Map reg val

Returns an initial state with all known registered set to a constant value.

exec :: (Ord reg, Num val) => Map reg val -> Instruction reg val -> Map reg val

Performs one operation on a register state, if the condition applies.

day8a :: String -> Int

day8b :: String -> Int