Day10

Description

Documentation

roll :: Int -> [a] -> (Int, Int) -> [a]

roll len list (count, skip) reverses the first count elements of list then rolls both those reversed elements and an additional skip elements from the start of the list to the end of the list.

len must be equal to length list.

rolls :: Int -> [Int] -> [Int]

rolls len counts iterates roll len list (count, skip), starting with list = [0..len-1], for each count in counts and skip = 0, 1, ... Finally, it unrotates by the total count + skip sum.

day10a :: Int -> String -> Int

day10b :: String -> String