Day15

Description

Documentation

parse :: String -> (Int, Int)

Returns the initial values for generators A and B.

genA :: Int -> Int

One step of generator A.

genB :: Int -> Int

One step of generator B.

genA' :: Int -> Int

Step generator A until a multiple of 4.

genB' :: Int -> Int

Step generator A until a multiple of 8.

gen :: (Int, Int) -> (Int, Int)

One step of both generators A and B.

gen' :: (Int, Int) -> (Int, Int)

Step both generators A and B until a multiple of 4 and 8 respectively.

count :: (a -> Bool) -> (a -> a) -> a -> Int -> Int

count p f x n == length (filter p . take n . tail $ iterate f x)

day15a :: String -> Int

day15b :: String -> Int