Day25

Description

Documentation

data State

The state of the the Turing machine.

Constructors

A

B

C

D

E

F

data Program state value

The setup for a run of a Turing machine.

Constructors

Program

Fields

start :: state

Initial state of the Turing machine.

steps :: Int

Number of cycles to run the Turing machine.

table :: Table Array state value

State transitions.

parseProgram :: (Ix state, Read state, Ix value, Num value, Read value) => ReadP (Program state value)

Parses a description to a Program.

day25 :: String -> IO Int